requirements.txt
google-cloud-bigquery>=2.34.4
google-cloud-storage>=2.7.0
google-cloud-aiplatform>=1.36.0  # For Vertex AI
google-auth>=2.6.0
vertexai>=0.5.0                  # Updated Vertex AI SDK
astor>=0.8.1
torch>=1.10.0                    # For PyTorch documentation
tensorflow>=2.8.0                # For TensorFlow documentation
scikit-learn>=1.0.0              # For scikit-learn documentation
pandas>=1.3.5                    # For data preparation
Flask>=2.0.1                     # For API service
gunicorn>=20.1.0                 # For serving the API
flask-cors>=3.0.10               # For CORS support in API